+++ application/views/actor.php
@@ -1,8 +1,13 @@
 <div class="row">
         <div class="span4">
                 <a href="/welcome/show/{key}">
-                        <img src="/pix/{who}.jpg" title="{who} plays as {role}"/>
-                </a>
-            
+                        <img src="/pix/{who}.jpg" title="{who} plays {role}"/>
+                </a>   
         </div>
+    
+    <h2>{who}</h2>
+        <?php
+            echo "I play as " . "{role}" . "<br>";
+            echo "Here is a quote I said: " . "{what}";
+        ?>
 </div>